HZ-0241 Move UI to public directory
